move to auth module, full working flow
